Mark external resource so Mix knows of dependency
By using `@external_resource`, Mix will know that the module depends on an external file. Changes to the contents of that file should then trigger a recompilation of the module. See https://hexdocs.pm/elixir/Module.html#module-external_resource

This is just a random contribution because I happened to read the code after seeing a tweet, and I maintain a library that also reads in a file at compile time. :wave: